Added location funtionality for individual entries

+++ Biodiversity/Biodiversity/Models/Animal.cs
@@ -10,9 +10,9 @@ namespace Biodiversity.Models
 
         public string Name { get; set; }
 
-        public float Lat { get; set; }
+        public double Lat { get; set; }
 
-        public float Lng { get; set; }
+        public double Lng { get; set; }
 
         public string Desc { get; set; }
